Replaced driver id int with long type

+++ src/main/java/com/test/example/repository/DriverRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DriverRepository extends PagingAndSortingRepository<Driver, Integer> {
+public interface DriverRepository extends PagingAndSortingRepository<Driver, Long> {
 
 }
